Revert "[Feat] better custom (#1027)"
This reverts commit 9e01aa0e0d9454ebf8c8703958577d3a177c0926.
