+++ src/prime_rl/trainer/custom_models/__init__.py
@@ -7,18 +7,15 @@ from transformers.models.auto.auto_factory import _BaseAutoModelClass, _LazyAuto
 from transformers.models.auto.configuration_auto import CONFIG_MAPPING_NAMES
 from transformers.models.llama.configuration_llama import LlamaConfig
 
-from prime_rl.trainer.models.glm import Glm4MoeConfig, Glm4MoeForCausalLM
-from prime_rl.trainer.models.llama import LlamaForCausalLM
-from prime_rl.trainer.models.qwen import Qwen3MoeConfig, Qwen3MoeForCausalLM
+from prime_rl.trainer.custom_models.glm import Glm4MoeConfig, Glm4MoeForCausalLM
+from prime_rl.trainer.custom_models.llama import LlamaForCausalLM
 
 # Make custom config discoverable by AutoConfig
 AutoConfig.register("glm4_moe", Glm4MoeConfig, exist_ok=True)
-AutoConfig.register("qwen3_moe", Qwen3MoeConfig, exist_ok=True)
 
 _CUSTOM_CAUSAL_LM_MAPPING = _LazyAutoMapping(CONFIG_MAPPING_NAMES, OrderedDict())
 _CUSTOM_CAUSAL_LM_MAPPING.register(LlamaConfig, LlamaForCausalLM, exist_ok=True)
 _CUSTOM_CAUSAL_LM_MAPPING.register(Glm4MoeConfig, Glm4MoeForCausalLM, exist_ok=True)
-_CUSTOM_CAUSAL_LM_MAPPING.register(Qwen3MoeConfig, Qwen3MoeForCausalLM, exist_ok=True)
 
 
 class AutoModelForCausalLMPrimeRL(_BaseAutoModelClass):
